Refactor: #225 CalendarViewModel 제거 및 각 상세화면의 뷰모델로 이관
