+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/homedetail/medicine/viewmodel/MedicineViewModel.kt
@@ -10,14 +10,35 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import org.koin.android.annotation.KoinViewModel
 import retrofit2.HttpException
+import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
+import java.time.temporal.TemporalAdjusters
 
 @KoinViewModel
 class MedicineViewModel(
     private val medicineRepository: MedicineRepository,
 ) : ViewModel() {
 
+    // 캘린더 상태
+    private val _selectedDate = MutableStateFlow(LocalDate.now())
+    val selectedDate: StateFlow<LocalDate> = _selectedDate
+    fun selectDate(date: LocalDate) {
+        _selectedDate.value = date
+    }
+
+    fun resetToToday() {
+        _selectedDate.value = LocalDate.now()
+    }
+
+    fun getCurrentWeekDates(): List<LocalDate> {
+        val base = _selectedDate.value
+        val startOfWeek =
+            base.with(TemporalAdjusters.previousOrSame(DayOfWeek.SUNDAY))
+        return (0..6).map { startOfWeek.plusDays(it.toLong()) }
+    }
+
+    // 복약 상태
     private companion object {
         const val TAG = "MED_API"
     }
